+++ react-components/src/components/ProductFinder/ProductFinder.jsx
@@ -1,71 +1,22 @@
-import React, { useState, useEffect } from 'react'
+/* eslint-disable prefer-destructuring */
+import React, { useState } from 'react'
 import ReactDOM from 'react-dom'
+import PropTypes from 'prop-types'
 import ReactModal from 'react-modal'
-import { getModalIsOpen, getProductsExpertise } from '@src/reducers'
 import Services from '@src/Services'
 import Spinner from '../Spinner/Spinner'
-import MessageConfirmation from './MessageConfirmation'
+import Confirmation from './MessageConfirmation'
 
-const customStyles = {
-  content: {
-    top: '0',
-    right: '0',
-    bottom: '0',
-    left: 'auto',
-    minWidth: '800px',
-    padding: '0',
-    border: 'none',
-    overflow: 'none'
-  },
-  overlay: {
-    background: 'rgb(45 45 45 / 45%)',
-    zIndex: '3'
-  }
-}
-
-function ValueChooser(attribute, handleChange) {
-  const changeValue = (element) => {
-    console.log(combinationValue) // TODO WIP
-  }
-  let startCombinationValue = {}
-  let profile = (attribute.attrs || []).map((option, index) => {
-    startCombinationValue[option.id] = option.value
-    return (
-      <label key={option.id} htmlFor={option.id} className="p-f-m m-b-xxs grid">
-        <div className="c-1-4">
-          <input
-            type="number"
-            className="form-control"
-            id={option.id}
-            name={attribute.id}
-            defaultValue={option.value}
-            data-label={option.name}
-            onChange={changeValue}
-          />
-        </div>
-        <div className="c-3-4">{option.name}</div>
-      </label>
-    )
-  })
-
-  return (
-    <div>
-      {profile}
-      <button className="button button--primary" onClick={handleChange}>
-        Send
-      </button>
-    </div>
-  )
-}
-
-export function ProductFinder(props) {
-  var searchInput
-  const [modalIsOpen, setIsOpen] = React.useState(false)
-  const [selectedProduct, setSelectedProduct] = React.useState(props.text)
-  const [searchResults, setSearchResults] = React.useState([])
-  const [isLoading, setLoading] = React.useState(false)
-  const [isScrolled, setIsScrolled] = React.useState(false)
-  const [productConfirmationRequired, setProductConfirmationRequired] = React.useState(false)
+function ProductFinder(props) {
+  const { text } = props;
+  let searchInput
+  const [modalIsOpen, setIsOpen] = useState(false)
+  const [selectedProduct, setSelectedProduct] = useState(text)
+  const [searchResults, setSearchResults] = useState([])
+  const [isLoading, setLoading] = useState(false)
+  const [isScrolled, setIsScrolled] = useState(false)
+  const [searchEnabled, setSearchEnabled] = useState(false)
+  const [productConfirmationRequired, setProductConfirmationRequired] = useState(false)
 
   const openModal = () => {
     setProductConfirmationRequired(!!selectedProduct)
@@ -84,18 +35,16 @@ export function ProductFinder(props) {
 
   const saveProduct = () => {
     setSelectedProduct(searchResults.currentItemName)
-    let result = Services.updateExportPlan({
-      export_commodity_codes: [
-        {
+    Services.updateExportPlan({
+        export_commodity_codes: [{
           commodity_name: searchResults.currentItemName,
           commodity_code: searchResults.hsCode
-        }
-      ]
-    })
-      .then((result) => {
+        }]
+      })
+      .then(() => {
         closeModal()
       })
-      .catch((result) => {
+      .catch(() => {
         // TODO: add an error dialogue here
       })
   }
@@ -104,25 +53,15 @@ export function ProductFinder(props) {
     searchInput.focus()
   }
 
-  const inputKeypress = (evt) => {
-    if (evt.key == 'Enter') {
-      evt.preventDefault()
-      search()
-    }
-  }
-
-  const onScroll = (evt) => {
-    setIsScrolled(evt.target.scrollTop > 0)
-  }
-
   const processResponse = (request) => {
     setLoading(true)
     request
       .then((result) => {
         setLoading(false)
-        console.log('Initial search result', result) // TODO: Needed during development
+        /* eslint-disable no-console */
+        console.log('Search result', result) // TODO: Needed during development
+        /* eslint-enable no-console */
         if (result && result.data && result.data.txId) {
-          console.log('set search result')
           setSearchResults(result.data)
         } else {
           setSearchResults(searchResults) // force re-render to reset any changed selectors
@@ -130,17 +69,43 @@ export function ProductFinder(props) {
       })
       .catch(() => {
         setLoading(false)
-        setSearchResults(result || {})
+        setSearchResults({})
       })
   }
 
   const search = () => {
-    let query = searchInput.value
-    processResponse(Services.lookupProduct({ q: query }))
+    const query = searchInput.value
+    if (query) {
+      processResponse(Services.lookupProduct({ q: query }))
+    }
   }
 
+  const inputKeypress = (evt) => {
+    if (evt.key === 'Enter') {
+      evt.preventDefault()
+      search()
+    }
+  }
+
+  const inputChange = (evt) => {
+    const value = evt.target.value
+    setSearchEnabled(!!value)
+  }
+
+  const clearSearchInput = (evt) => {
+    const input = evt.target.parentElement.querySelector('input')
+    input.value = ''
+    input.focus()
+    setSearchEnabled(false)
+  }
+
+  const onScroll = (evt) => {
+    setIsScrolled(evt.target.scrollTop > 0)
+  }
+
+
   const RadioButtons = (attribute, handleChange, setValue = true) => {
-    let buttons = (attribute.attrs || []).map((option, index) => {
+    const buttons = (attribute.attrs || []).map((option) => {
       return (
         <label key={option.id} htmlFor={option.id} className="multiple-choice p-f-m m-b-xxs">
           <input
@@ -150,10 +115,13 @@ export function ProductFinder(props) {
             name={attribute.id}
             value={option.id}
             data-label={option.name}
-            defaultChecked={setValue && option.value == 'true'}
+            defaultChecked={setValue && option.value === 'true'}
+            aria-label={option.name}
           />
           {option.name}
-          <label htmlFor={option.id}></label>
+          {/* eslint-disable jsx-a11y/label-has-associated-control */}
+          <label htmlFor={option.id}/>
+          {/* eslint-enable jsx-a11y/label-has-associated-control */}
         </label>
       )
     })
@@ -176,7 +144,7 @@ export function ProductFinder(props) {
       }
     }
 
-    let body = { SELECTION: RadioButtons, VALUED: ValueChooser }[attribute.type](
+    const body = { SELECTION: RadioButtons, VALUED: RadioButtons } [attribute.type](
       attribute,
       handleChange,
       !section.isItemChoice
@@ -191,12 +159,12 @@ export function ProductFinder(props) {
   }
 
   const Section = (title, sectionDetails) => {
-    if (!sectionDetails || sectionDetails.length == 0 || !sectionDetails.map) return null
+    if (!sectionDetails || sectionDetails.length === 0 || !sectionDetails.map) return null
     return (
       <section className="summary">
         <h3 className="h-s p-0">{title}</h3>
         <div className="">
-          {(sectionDetails || []).map((value, index) => {
+          {(sectionDetails || []).map((value) => {
             return Attribute(value, sectionDetails)
           })}
         </div>
@@ -204,30 +172,28 @@ export function ProductFinder(props) {
     )
   }
 
-  const buildMap = (block, map) => {
-    // build an intetrraction block, removing any duplicates from previous
-    let newBlock = []
-    for (var index in block) {
-      let interraction = block[index]
-      if (interraction && interraction.id) {
-        if (!map[interraction.id]) {
-          map[interraction.id] = true
-          newBlock.push(interraction)
+  const buildMap = (block) => {
+    // build an interaction block, removing any duplicates from previous
+    const newBlock = []
+    if (block && block.length) {
+      for (let index = 0; index < block.length; index += 1) {
+        const interaction = block[index]
+        if (interaction && interaction.id) {
+          newBlock.push(interaction)
         }
       }
     }
     return newBlock.length ? newBlock : null
   }
 
-  const resultsDisplay = (searchResults) => {
-    // Build maps of intteractions as we don't want any duplicates
-    let iMap = {}
-    let questions = buildMap([searchResults.currentQuestionInteraction], iMap)
-    let assumptions = buildMap(searchResults.assumedInteractions, iMap)
-    let known = buildMap(searchResults.knownInteractions, iMap)
-    let itemChoice = buildMap([searchResults.currentItemInteraction], iMap)
-    ;(itemChoice || {}).isItemChoice = true
-    let spinner = isLoading ? (
+  const resultsDisplay = (results) => {
+    // Build maps of interactions as we don't want any duplicates
+    const questions = buildMap([results.currentQuestionInteraction])
+    const assumptions = buildMap(results.assumedInteractions)
+    const known = buildMap(results.knownInteractions)
+    const itemChoice = buildMap([results.currentItemInteraction]);
+    (itemChoice || {}).isItemChoice = true
+    const spinner = isLoading ? (
       <div className="shim">
         <Spinner text="" />
       </div>
@@ -246,7 +212,7 @@ export function ProductFinder(props) {
           {searchResults.hsCode && (
             <section className="found-section grid bg-black-10">
               <div className="c-1-3">
-                <span className="h-s">You've found your product!</span>
+                <span className="h-s">You&apos;ve found your product!</span>
               </div>
               <div className="c-1-3">
                 <div className="h-xs p-t-0 capitalize">{searchResults.currentItemName}</div>
@@ -262,53 +228,65 @@ export function ProductFinder(props) {
           {false && searchResults.productDescription && (
             <section className="summary table">
               <div className="table-row">
-                <div className="table-cell">Here's what we know about your</div>
+                <div className="table-cell">Here&apos;s what we know about your</div>
                 <div className="table-cell bold capitalize">{searchResults.productDescription}</div>
               </div>
             </section>
           )}
           {Section('Please choose your item', itemChoice)}
           {!itemChoice && Section(`Tell us more about your '${searchResults.currentItemName}'`, questions)}
-          {!itemChoice && Section(`Your item's characteristics`, known)}
-          {!itemChoice && Section("We've assumed:", assumptions)}
+          {!itemChoice && Section('Your item&apos;s characteristics', known)}
+          {!itemChoice && Section('We\'ve assumed:', assumptions)}
         </div>
       </div>
     )
   }
 
-  let buttonClass = 'tag ' + (!selectedProduct ? 'tag--tertiary' : '') + ' tag--icon'
-  let scrollerClass = 'scroll-area ' + (isScrolled ? 'scrolled' : '')
+  const buttonClass = `tag ${!selectedProduct ? 'tag--tertiary' : ''} tag--icon`
+  const scrollerClass = `scroll-area ${isScrolled ? 'scrolled' : ''}`
+  const headerHeight = '190px'
 
   return (
     <span>
-      {/* eslint-disable-next-line react/button-has-type */}
-      <button className={buttonClass} onClick={openModal}>
+      <button type="button" className={buttonClass} onClick={openModal}>
         {selectedProduct || 'add product'}
-        <i className="fas fa-chevron-right"></i>
+        <i className={`fa ${selectedProduct ? 'fa-edit' : 'fa-plus'}`}/>
       </button>
-      <ReactModal isOpen={modalIsOpen} onRequestClose={closeModal} style={customStyles} onAfterOpen={modalAfterOpen}>
+      <ReactModal 
+        isOpen={modalIsOpen} 
+        onRequestClose={closeModal} 
+        className="modal max-modal p-v-s p-h-l"
+        overlayClassName="modal-overlay center"
+        onAfterOpen={modalAfterOpen}
+      >
         <form className="product-finder">
-          <div className="search-header bg-blue-deep-80 text-white p-s" style={{ height: '172px' }}>
-            <button className="pull-right m-r-0 dialog-close" onClick={closeModal}></button>
-            <h3 className="h-s text-white p-t-0">Search by name</h3>
+          <div className="modal-header" style={{height:headerHeight}}>        
+            <button id="dialog-close" type="button" aria-label="Save" className="pull-right m-r-0 dialog-close" onClick={closeModal}/>
+            <h3 className="h-m p-t-0">Search by name</h3>
             <div>Find the product you want to export</div>
-            <input
-              className="form-control c-2-3"
-              type="text"
-              ref={(_searchInput) => (searchInput = _searchInput)}
-              onKeyPress={inputKeypress}
-              defaultValue=""
-            />
-            <button className="button button--tertiary m-f-xxs" type="button" onClick={search}>
-              Search
-            </button>
+            <div className="flex-centre m-t-xs search-input">
+              <div className="flex-centre">
+                <input
+                  className="form-control"
+                  type="text"
+                  ref={(_searchInput) => {searchInput = _searchInput}}
+                  onKeyPress={inputKeypress}
+                  onChange={inputChange}
+                  defaultValue=""
+                />
+                <button type="button" aria-label="Clear" className="fa fa-times clear" onClick={clearSearchInput}/>
+                </div>
+              <button className="button button--small button--only-icon m-f-xs" disabled={!searchEnabled} type="button" onClick={search}>
+                <i className="fa fa-arrow-right"/>
+              </button>
+            </div>
           </div>
-          <div className={scrollerClass} style={{ marginTop: '172px' }} onScroll={onScroll}>
+          <div className={scrollerClass} style={{marginTop:headerHeight}} onScroll={onScroll}>
             {resultsDisplay(searchResults)}
           </div>
         </form>
       </ReactModal>
-      <MessageConfirmation
+      <Confirmation
         buttonClass={buttonClass}
         productConfirmation={productConfirmationRequired}
         handleButtonClick={closeConfirmation}
@@ -320,10 +298,14 @@ export function ProductFinder(props) {
   )
 }
 
-export default function({ ...params }) {
+ProductFinder.propTypes = {
+  text: PropTypes.string.isRequired,
+}
+
+export default function createProductFinder({ ...params }) {
   const mainElement = document.createElement('span')
   document.body.appendChild(mainElement)
   ReactModal.setAppElement(mainElement)
-  let text = params.element.getAttribute('data-text')
-  ReactDOM.render(<ProductFinder text={text}></ProductFinder>, params.element)
+  const text = params.element.getAttribute('data-text')
+  ReactDOM.render(<ProductFinder text={text}/>, params.element)
 }