+++ CHANGELOG.md
@@ -58,6 +58,7 @@
 - GP2-315 - route to market backend integration
 
 ### Implemented enhancements
+- GP2-680 - Restyle product search modal
 - GP2-602 - Use new header on all pages and allow login/logout
 - GP2-331 - Rerouting to login page with error message
 - GP2-413 -  style learn linking component